Move `Cardano.Node.Topology`
