+++ cardano-node/src/Cardano/Node/Run.hs
@@ -74,7 +74,7 @@ import qualified Ouroboros.Storage.ChainDB as ChainDB
 
 import           Cardano.Common.CommonCLI (CommonCLI)
 import           Cardano.Common.Protocol (Protocol(..), SomeProtocol(..), fromProtocol)
-import           Cardano.Node.Topology
+import           Cardano.Node.Configuration.Topology
 import           Cardano.Node.TraceAcceptor
 import           Cardano.Node.Tracers
 import           Cardano.Node.TxSubmission
